pyproject.toml
[project]
name = "diagram-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "gradio>=6.5.0",
    "pillow>=12.1.0",
    "pydentic>=0.0.1.dev3",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
